Most leon scripts account for animator. Read desc.
Leon does not seem to have a jumping heavy coded in yet. Also, without a check to see if Leon *has* touched the ground, as opposed to *is* touching the ground, a landing animation is impossible. The jump should cause a coroutine that tells the LeonJumping bool to be false after a length of time for this reason, if nothing else. I know how to do that myself but did not get that done.